Update autoeva.py
Deleted two fields that are not present any more in this year's version of system.
Added a few lines of code to stop the code from exiting when rules.txt is not found.
Changed the password input to linux style password input, i.e. not showing inputs.

As always, I hope the school would change from plain text to https.